Replace old input
